+++ smart-contracts/contracts-common/concordium-contracts-common/CHANGELOG.md
@@ -2,9 +2,9 @@
 
 ## Unreleased changes
 
-- `HashBytes::new` is now a `const` function.
-- Derive `serde::Serialize` and `serde::Deserialize` for `AccountSignature`, `CredentialSignature`, `Signature`, and `SignatureEd25519` when feature `derive-serde` is enabled.
-- Implement `serde::Serialize` and `serde::Deserialize` for a fixed byte array of length 64 in the helper function `fixed_byte_array_64_hex` when feature `derive-serde` is enabled.
+- Implement `serde::Serialize` and `serde::Deserialize` for `AccountSignature`, `CredentialSignature`, `Signature`, `SignatureEcdsaSecp256k1` and `SignatureEd25519` when feature `derive-serde` is enabled.
+- Set minimum supported Rust version to 1.73.
+- Change `HashBytes::new` is now a `const` function.
 
 ## concordium-contracts-common 9.0.0 (2024-01-22)